Added Wall Slide
And some optimization
