+++ Assets/_Project/RGScripts/_Refactored/Player/PlayerStates/IdleState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1bc352f7096f48d99cfb25ff3f545051
+timeCreated: 1730270205
